api/requirements.txt
# Django API Dependencies - Passwordless Authentication
# Django 6.0 requires Python 3.12+

# Core Django - pin specific versions to avoid resolution issues
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0
django-environ==0.12.0
dj-database-url==2.3.0
django-unfold==0.75.0

# Authentication
django-allauth==65.13.1
django-rest-knox==5.0.2
django-otp==1.5.4
firebase-admin==6.6.0
qrcode==8.0
webauthn==2.2.0
fido2==1.2.0

# Database
psycopg2-binary==2.9.10
redis==5.2.1
django-redis==5.4.0

# WebSocket Support
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Background Tasks
celery==5.4.0
# django-celery-beat Django 6.0 support merged but not released yet (PR #978)
# Install from git with Django 6.0 support commit
django-celery-beat @ git+https://github.com/celery/django-celery-beat@bd429063cbb227c00905e7693c38895e44ff8e47

# NATS Messaging (alerts lifecycle + NLP progress)
nats-py>=2.9.0

# Security
cryptography==44.0.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2
croniter==5.0.1
structlog==24.4.0
rich==13.9.4

# Testing dependencies
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==6.0.0
pytest-xdist==3.6.1
testcontainers[postgres,redis,nats,compose]==4.10.0
factory-boy==3.3.1
faker==33.1.0
freezegun==1.5.1
responses==0.25.6

# Production WSGI Server
gunicorn==23.0.0
whitenoise==6.8.2

# Development
black==24.10.0
isort==5.13.2
mypy==1.13.0
django-extensions==3.2.3
ipython==8.31.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# API Documentation
drf-spectacular==0.28.0

# Monitoring
django-prometheus==2.3.1
sentry-sdk==2.19.2

# Data Processing
pandas==2.2.3
numpy==2.2.1
resend==2.7.0

# DuckLake Analytics (requires DuckDB 1.3.0+ for ducklake extension)
duckdb>=1.3.0
pyarrow>=15.0.0

# LLM Integration (for NLP alert processing)
# Direct LiteLLM calls with CoT prompting for NLP
litellm==1.80.16
openai==2.15.0  # Required by litellm and DSPy
google-generativeai>=0.8.0  # Required for Gemini via LiteLLM

# DSPy (prompt compiler + optimization loop)
dspy==3.1.0
